==> RchaeoStats-workshop/scripts/02_data-transform.R <==
library(tidyverse)
library(here)

mortuary_data <- read_csv(here("data/mortuary_clean.csv"))


# Subsetting data ---------------------------------------------------------

# select and filter

select(mortuary_data, Golden_bead) # select by column name
select(mortuary_data, Golden_bead, Glass_bead) # multiple colunms
select(mortuary_data, 22, 23) # position

filter(mortuary_data, Phase == "pre") # filter the rows where Phase is 'pre'
filter(mortuary_data, start_layer == 5)
filter(mortuary_data, start_layer != 5)
filter(mortuary_data, start_layer > 5) 
filter(mortuary_data, start_layer > 5 & start_layer <= 12)
filter(mortuary_data, start_layer > 5 & Layer == 1)
filter(mortuary_data, start_layer > 5 | start_layer < 12)
filter(mortuary_data, is.na(start_layer)) # filter the rows where Phase is missing (NA)
filter(mortuary_data, !is.na(start_layer)) # filter the rows where Phase is missing (NA)

select(filter(mortuary_data, Phase == "pre"), Phase)
# nested functions
select(filter(mortuary_data, Phase == "pre"), ID, Golden_bead, Glass_bead)
# or create intermediate objects
pre_euro <-filter(mortuary_data, Phase == "pre")
select(pre_euro, Golden_bead, Glass_bead)  

pre_condition <- filter(mortuary_data, Phase == "pre", Condition == 2)
mean(pre_condition$Golden_bead, na.rm = TRUE)
  
  

# Groups ------------------------------------------------------------------


group_by(.data = mortuary_data, Phase)
class(group_by(.data = mortuary_data, Phase))

summarise(
  mortuary_data,
  mean_golden = mean(Golden_bead, na.rm = TRUE),
  median_golden = median(Golden_bead, na.rm = TRUE),
  sd_golden = sd(Golden_bead, na.rm = TRUE)
)

mortuary_data |>
  group_by(Phase) |>
  summarise(
    mean_golden = mean(Golden_bead, na.rm = TRUE),
    median_golden = median(Golden_bead, na.rm = TRUE),
    sd_golden = sd(Golden_bead, na.rm = TRUE)
  )
  
summarise(
  mortuary_data,
  mean_golden = mean(Golden_bead, na.rm = TRUE),
  median_golden = median(Golden_bead, na.rm = TRUE),
  sd_golden = sd(Golden_bead, na.rm = TRUE),
  .by = Phase
) 

# sample/group size
mortuary_data |>
  group_by(Phase) |>
  summarise(n = n())

mortuary_data |>
  count(Phase)


# Across ------------------------------------------------------------------

# The same operation/functino across multiple columns

mortuary_data |>
  group_by(Phase) |>
  summarise(
    across(c(Golden_bead, Glass_bead, IndoPacific_bead), ~ mean(.x, na.rm = T))
  )

mortuary_data |>
  group_by(Phase) |>
  summarise(
    across(Agate_bead:Metal_piece, ~ mean(.x, na.rm = T))
  )

# Selection helpers
mortuary_data |>
  group_by(Phase) |>
  summarise(
    across(contains("bead"), ~ mean(.x, na.rm = T))
  )

mortuary_data |>
  group_by(Phase) |>
  summarise(
    across(ends_with("_bead"), ~ mean(.x, na.rm = T))
  )

mortuary_data |>
  select(ends_with("bead"))
  

## Exercise
# select all columns with a 'Metal' artefact, then calculate the mean count for each Phase

mortuary_data |>
  group_by(Phase) |> 
  summarise(
    across(
      contains("Metal"),
      ~ mean(.x, na.rm = T)
    )
  )

## End of exercise

# multiple operations/functions across

mortuary_data |>
  group_by(Phase) |>
  summarise(
    across(
      #everything(),
      ends_with("bead"),
      list(
        mean = ~ mean(.x, na.rm = T),
        sd = ~ sd(.x, na.rm = T)
      )
    )
  )

my_list <- list(
  "a_vector" = mortuary_data$Width,
  "a_data_frame" = mortuary_data,
  "a_function" = mean,
  "etc" = "etc"
)

my_list$a_vector
my_list$a_data_frame$Layer[3] # third item of the Layer variable from the data frame from the list
my_list[1] # first item of the list
my_list[[1]] # first item of the list in its actual original object


# Pivoting ----------------------------------------------------------------

artefact_long <- mortuary_data |>
  select(ID, Phase, Agate_bead, Golden_bead) |>
  pivot_longer(Agate_bead:Golden_bead)

artefact_long$value

artefact_longer <- mortuary_data |>
  pivot_longer(
    Agate_bead:Metal_piece,
    names_to = "artefact",
    values_to = "count"
  )
dim(artefact_longer)


artefact_longer |>
  group_by(artefact) |>
  summarise(
    mean = mean(count, na.rm = T),
    sd = sd(count, na.rm = T)
  )
mortuary_data$Knit

# makes the artefacts easier to plot because we have a single column for artefact name and single column for the counts
artefact_longer |>
  ggplot(aes(x = artefact, y = count)) +
  geom_col()

artefact_longer |>
  filter(
    artefact != "IndoPacific_bead",
#    !is.na(Phase) # can use this instead of remove_missing(vars = "Phase")
    ) |>
  remove_missing(vars = "Phase") |>
  ggplot(aes(x = artefact, y = count, fill = Phase)) +
  geom_col() +
  scale_fill_viridis_d(option = "plasma")

percent_artefact <- artefact_longer |>
  group_by(Phase, artefact) |>
  summarise(n = sum(count, na.rm = TRUE)) |>
  group_by(Phase) |>
  mutate(percent = (n / sum(n)) * 100)

percent_artefact |>
  ggplot(aes(x = Phase, y = percent, fill = artefact)) +
    geom_col()


artefact_longer |>
  remove_missing(vars = "Phase") |>
  ggplot(aes(x = Phase, y = count, fill = Phase)) +
  geom_col() +
  facet_wrap(~ artefact, scales = "free_y")

artefact_longer |>
  filter(str_detect(artefact, "bead")) |>
  ggplot(aes(x = artefact, y = count)) +
  geom_col()

str_detect(artefact_longer$artefact, "bead")

total_artefact_count <- artefact_longer |>
  group_by(Phase, artefact) |>
  summarise(
    total_count = sum(count, na.rm = T)
  ) |>
  ungroup()

write_csv(total_artefact_count, here("data/total-artefact-counts.csv"))

artefact_wide <- artefact_longer |>
  pivot_wider(names_from = artefact, values_from = count)

dim(artefact_wide) == dim(mortuary_data)
